pyproject.toml
[project]
name = "llm-wiki"
version = "0.1.0"
description = "Personal knowledge base engine (LlamaIndex + Chroma + llama.cpp + MCP)"
requires-python = ">=3.11"
dependencies = [
    "llama-index-core>=0.12.0",
    "llama-index-vector-stores-chroma>=0.4.0",
    "chromadb>=0.5.20",
    "httpx>=0.27.0",
    "fastmcp>=0.2.0",
    "typer>=0.12.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
llm-wiki = "llm_wiki.cli:app"
llm-wiki-mcp = "llm_wiki.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_wiki"]

[tool.ruff]
line-length = 100
target-version = "py311"
